Some renaming for clarity
The modifications here are aimed to remove reference to "double calls" since
this PR enables up to 8 additional calls.  They are also aimed to make it clear
that these radiation calls do not affect the evolution of the simulation, and
that the diagnostics pertain to what we get when we scale CO2 (as opposed to
some other constituent of the atmosphere).